requirements.txt
selenium
black
isort
flake8
mypy
pytest
pytest-cov
python-dotenv
price-parser
sqlalchemy
psycopg2
uuid
boto3
-e .